Make PU/QU etc. doubles rather than qint32s.
This is the hottest loop in the code, and profiling with "perf annotate"
showed that the double to integer conversions were the most expensive
instructions. These values are only going to be converted back to
doubles later anyway, so we may as well keep them as doubles.

About 18% speedup with -t1.

+++ tools/ld-comb-pal/palcolour.cpp
@@ -249,7 +249,7 @@ QByteArray PalColour::performDecode(QByteArray firstFieldData, QByteArray second
                 // NB: Multiline averaging/filtering assumes perfect
                 //     inter-line phase registration...
 
-                qint32 PU,QU, PV,QV, PY,QY;
+                double PU,QU, PV,QV, PY,QY;
                 for (qint32 i = videoParameters.activeVideoStart; i < videoParameters.activeVideoEnd; i++) {
                     PU=QU=0; PV=QV=0; PY=QY=0;
